remove CaptainService for refactor.
